- Fixed some warnings about float to int conversions - Changed TargetAction to use an instance of the destination class. - Fixed cyclic include in Destination and GameEntity. - Added prototype for eating action into Animal.

+++ simplerpg/Game.cpp
@@ -273,8 +273,8 @@ EntityList Game::getUnderCursor()
 	{
 		GameEntity *entity = *iter;
 		Vector2f pos = entity->getPosition();
-		int posX = round(pos.x);
-		int posY = round(pos.y);
+		int posX = (int)round(pos.x);
+		int posY = (int)round(pos.y);
 		if(posX == mCursorX && posY == mCursorY)
 		{
 			mUnderCursor.push_back(entity);
